requirements.txt
requests
google-cloud-bigquery
python-dotenv
flask
functions-framework
